Remove dependency from context to assert package
