--- context/export_test.go
@@ -1,5 +1,0 @@
-package context
-
-func ContainsAssertion() interface{} {
-	return assertions["contains"]
-}
